# src/main/java/org/openapitools/api/SigningBasketsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ErrorResponse;
import org.openapitools.model.SigningBasketStatusResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-13T04:23:17.978812429Z[Etc/UTC]", comments = "Generator version: 7.20.0-SNAPSHOT")
@Validated
@Tag(name = "Signing Baskets", description = "APIs for signing-baskets management")
public interface SigningBasketsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_SIGNING_BASKET_STATUS = "/v1/signing-baskets/{basketId}/status";
    /**
     * GET /v1/signing-baskets/{basketId}/status : Get signing basket status
     *
     * @param basketId Identifier of the signing basket (required)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Not Found (status code 404)
     *         or Server Error (status code 500)
     */
    @Operation(
        operationId = "getSigningBasketStatus",
        summary = "Get signing basket status",
        tags = { "Signing Baskets" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SigningBasketStatusResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "OAuth2"),
            @SecurityRequirement(name = "BasicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = SigningBasketsApi.PATH_GET_SIGNING_BASKET_STATUS,
        produces = { "application/json" }
    )
    default ResponseEntity<SigningBasketStatusResponse> getSigningBasketStatus(
        @Parameter(name = "basketId", description = "Identifier of the signing basket", required = true, in = ParameterIn.PATH) @PathVariable("basketId") String basketId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"basketId\" : \"3dc3d5b3-7023-4848-9853-f5400a64e80f\", \"transactionStatus\" : \"ACTC\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}