# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    name: WSO2
    url: https://wso2.com/solutions/financial-services/open-banking/
  description: This API defines the REST API contract for services that implements
    logic to extend the Open Data accelerator flow.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: API contract for financial accelerator extension points in WSO2 IS and APIM
  version: v1.0.3
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/wso2-f5b/OB4/1.0.0
security:
- BasicAuth: []
- OAuth2: []
...
paths:
...
  /v1/signing-baskets/{basketId}/status:
    get:
      operationId: getSigningBasketStatus
      parameters:
      - description: Identifier of the signing basket
        example: 3dc3d5b3-7023-4848-9853-f5400a64e80f
        explode: false
        in: path
        name: basketId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                basketId: 3dc3d5b3-7023-4848-9853-f5400a64e80f
                transactionStatus: ACTC
              schema:
                $ref: "#/components/schemas/SigningBasketStatusResponse"
          description: OK
        "400":
          content:
            application/json:
              example:
                responseId: Ec1wMjmiG8
                status: ERROR
                errorMessage: invalid_request
                errorDescription: Invalid basket ID format
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              example:
                responseId: Ec1wMjmiG8
                status: ERROR
                errorMessage: not_found
                errorDescription: Signing basket not found
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              example:
                responseId: Ec1wMjmiG8
                status: ERROR
                errorMessage: server_error
                errorDescription: Failed to retrieve signing basket status
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Server Error
      summary: Get signing basket status
      tags:
      - Signing Baskets
      x-accepts:
      - application/json
      x-tags:
      - tag: Signing Baskets
components:
  schemas:
    SigningBasketStatusResponse:
      description: Response containing the status of a signing basket
      example:
        basketId: 3dc3d5b3-7023-4848-9853-f5400a64e80f
        transactionStatus: ACTC
      properties:
        basketId:
          description: Identifier of the signing basket
          type: string
        transactionStatus:
          description: The transaction status of the signing basket
          enum:
          - RCVD
          - PATC
          - ACTC
          - ACCP
          - ACSC
          - ACSP
          - PDNG
          - RJCT
          - CANC
          type: string
      required:
      - basketId
      - transactionStatus
      type: object
...
    ErrorResponse:
      example:
        data: "{}"
        status: ERROR
      properties:
        status:
          description: "Indicates the outcome of the request. For an error operation,\
            \ this should be set to ERROR."
          enum:
          - ERROR
          type: string
        data:
          description: :"Custom error object to response back"
          type: object
      required:
      - data
      - status
      type: object
...
  securitySchemes:
    BasicAuth:
      scheme: basic
      type: http
    OAuth2:
      flows:
        clientCredentials:
          scopes:
            process: process request generate response
          tokenUrl: https://example.com/oauth/token
      type: oauth2